kernel/x86_64 SMP: fix IDT off-by-one corrupting k_percpu_stacks (SYSCALL triple-fault)
cpunum=1 booted through VM and all servers, then silently triple-faulted on
userland's first SYSCALL-based IPC (libc_ipc=1, which forces the INT IPC
fallback, booted to login -- the decisive clue).

Root cause: lapic_set_dummy_handlers() (APIC_DEBUG only) wrote one IDT gate past
the 256-entry table. The lapic_intr_dummy_handles_start label was not aligned to
the handler stride (the .balign is inside each LAPIC_INTR_DUMMY_HANDLER macro),
so it landed ~30 bytes before handler 0; the loop's "handler < end" bound then
ran one extra iteration, calling int_gate_idt(256, ...). idt[256] is adjacent to
k_percpu_stacks[0] in BSS, so this corrupted the per-CPU kernel stack pointer
that the SYSCALL fast-path loads (the INT path uses tss[].rsp0 and was immune).

Fixes:
- apic_asm.S: .balign LAPIC_INTR_DUMMY_HANDLER_SIZE before the start label so it
  coincides with handler 0 (also fixes the installed gate offsets, which were
  30 bytes low).
- apic.c: bound the install loop with "&& vect < IDT_SIZE" (defensive).
- mpx.S: per-CPU SYSCALL stub runs RESTORE_KERNEL_SEGS before entering C,
  matching the INT entry path (SYSCALL leaves user data segments loaded).
- arch_proto.h: K_STACK_SIZE 4KB -> 16KB; it doubles as the cpuid-macro rounding
  granularity and x86_64 frames are ~2x i386's, so one page is too tight.

Also folded in the SMP x86_64 build/bring-up fixes from this hunt:
- Makefile.inc: assemble trampoline.S with clang's integrated assembler (the
  external x86_64-elf64-minix-as lacks the elf64-x86-64 BFD vector).
- (int)cpuid casts for -Werror=format (utility.c, apic.c, proc.c).
- Seed the boot-stack cpuid slot to the BSP id in smp_init()/arch_init() so
  pre-switch_k_stack code (LAPIC calibration, arch_post_init ptproc=vm) indexes
  per-CPU state correctly (x86_64 and i386).

docs/kernel-arch-x86_64.md: document the IDT/k_percpu_stacks adjacency bug and
the SYSCALL-path corrections.

+++ minix/kernel/proc.c
@@ -438,7 +438,7 @@ check_misc_flags:
 
 	TRACE(VF_SCHEDULING, printf("cpu %d starting %s / %d "
 				"pc 0x%08lx\n",
-		cpuid, p->p_name, p->p_endpoint, (unsigned long)p->p_reg.pc););
+		(int)cpuid, p->p_name, p->p_endpoint, (unsigned long)p->p_reg.pc););
 #if DEBUG_TRACE
 	p->p_schedules++;
 #endif
@@ -1810,7 +1810,7 @@ static struct proc * pick_proc(void)
   rdy_head = get_cpulocal_var(run_q_head);
   for (q=0; q < NR_SCHED_QUEUES; q++) {	
 	if(!(rp = rdy_head[q])) {
-		TRACE(VF_PICKPROC, printf("cpu %d queue %d empty\n", cpuid, q););
+		TRACE(VF_PICKPROC, printf("cpu %d queue %d empty\n", (int)cpuid, q););
 		continue;
 	}
 	assert(proc_is_runnable(rp));
